+++ .gitignore
@@ -3,9 +3,9 @@
 *.log
 *.pyc
 *.swp
-.DS_Store
-.atom/
-.buildlog/
+**/.DS_Store
+**/.atom/
+**/.buildlog/
 .history
 .svn/
 
@@ -13,7 +13,7 @@
 *.iml
 *.ipr
 *.iws
-.idea/
+**/.idea/
 
 # The .vscode folder contains launch configuration and tasks you configure in
 # VS Code which you may wish to be included in version control, so this line
@@ -23,24 +23,24 @@
 # Flutter/Dart/Pub related
 **/doc/api/
 **/ios/Flutter/.last_build_id
-.dart_tool/
-.flutter-plugins
-.flutter-plugins-dependencies
-.packages
-.pub-cache/
-.pub/
-/build/
+**/.dart_tool/
+**/.flutter-plugins
+**/.flutter-plugins-dependencies
+**/.packages
+**/.pub-cache/
+**/.pub/
+**/build/
 
 # Web related
-lib/generated_plugin_registrant.dart
+frontend/lib/generated_plugin_registrant.dart
 
 # Symbolication related
-app.*.symbols
+**/app.*.symbols
 
 # Obfuscation related
-app.*.map.json
+**/app.*.map.json
 
 # Android Studio will place build artifacts here
-/android/app/debug
-/android/app/profile
-/android/app/release
+frontend/android/app/debug
+frontend/android/app/profile
+frontend/android/app/release
